+++ paypal-user-service/src/main/java/com/amaris/it/paypal/user/repository/TransactionStatusNotifier.java
@@ -2,5 +2,5 @@ package com.amaris.it.paypal.user.repository;
 
 public interface TransactionStatusNotifier {
 
-  void notify(Integer transactionId, String status);
+  void notify(Long transactionId, String status);
 }
